Resolve Comments
