+++ ingestion/tests/unit/topology/database/test_datalake.py
@@ -31,7 +31,8 @@ from metadata.generated.schema.metadataIngestion.workflow import (
 )
 from metadata.generated.schema.type.entityReference import EntityReference
 from metadata.ingestion.source.database.datalake.metadata import DatalakeSource
-from metadata.utils.datalake_utils import read_from_avro, read_from_json
+from metadata.utils.datalake.avro_dispatch import read_from_avro
+from metadata.utils.datalake.json_dispatch import read_from_json
 
 mock_datalake_config = {
     "source": {
